Reintroduce initial item lifetime synchronization
- only sent if necessary
- also added the option to completely exclude an item from flower interactions by setting its lifetime to Short.MIN_VALUE

+++ Xplat/src/main/java/vazkii/botania/xplat/XplatAbstractions.java
@@ -153,7 +153,7 @@ public interface XplatAbstractions {
 	Optional<ItemSource> getItemSource(ItemEntity item);
 	void setItemSource(ItemEntity item, ItemSource source);
 
-	int getItemLifeTime(ItemEntity item);
+	short getItemLifeTime(ItemEntity item);
 	void setItemLifeTime(ItemEntity item, int ticks);
 
 	List<ItemStack> getKeptItems(Player player);
